Preventing weapons from dislodging in disabled victims
If a victim is stunned or otherwise disabled, there's no reason they should be able to trap your weapon in their body.  My plan is to come back and embellish on this (because you lose a lot of damage on crit hits that dislodge weapons that never get pulled out, and we should add that damage back in) and there are some skills checks which could come into play to add some neat flavor text (weapon retention, pierce prof/spec, telekinesis, etc.).